remove: Col-Type-Header; rename/move CRModal to 'assertions/'
