+++ static_report/src/components/shared/Charts/index.ts
@@ -2,5 +2,6 @@ export * from './BooleanPieChart';
 export * from './ChartContainer';
 export * from './FlatBoxPlotChart';
 export * from './FlatStackedBarChart';
+export * from './TopKSummaryList';
 export * from './HistogramChart';
 export * from './utils';
