+++ static_report/src/components/shared/Columns/index.ts
@@ -1,4 +1,3 @@
-export * from './ColumnTypeHeader';
 export * from './ColumnDetailMasterList';
 export * from './ColumnMetrics';
 export * from './constants';
